Run coveralls under Python 2.7 for convenience
Ubuntu Precise, the current Travis version, doesn't provide
`python3-pip`. As an experiment, and since this is "only" test coverage,
this drops coverage run to ("native") 2.7 rather than faff with getting
a Python 3 pip as "pip" or doing some sort of conditional hackery on the
"pip install" to use "pip-3.2" in some cases.